기업정보 - 기업회원 연관관계 매핑시 FK를 기원회원 PK로 변경

+++ JOBGO-COMMON/src/main/java/com/flab/jobgo/common/dto/EnterpriseInfoResponseDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class EnterpriseInfoResponseDTO {
 	
 	// 기업Id
-	private Integer enterpriseId;
+	private Long enterpriseId;
 	
 	// 기업명
 	private String enterpriseName;
@@ -55,7 +55,7 @@ public class EnterpriseInfoResponseDTO {
 	private Integer averageSalary;
 	
 	@Builder
-	public EnterpriseInfoResponseDTO(Integer enterpriseId, String enterpriseName, String companyNumber, String managerName, String contact,
+	public EnterpriseInfoResponseDTO(Long enterpriseId, String enterpriseName, String companyNumber, String managerName, String contact,
 			String fsIndustryCode, String fsIndustry, String scIndustryCode, String scIndustry, String address,
 			String establishDate, String enterpriseTypeCode, String enterpriseType, Integer employees,
 			Integer averageSalary) {
